% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simstar.env.R
\docType{data}
\name{.simstar.env}
\alias{.simstar.env}
\title{Package environment}
\format{An object of class \code{environment} of length 0.}
\usage{
.simstar.env
}
\description{
Environment used to store paths to external code.
}
\seealso{
\code{\link{paths}}
}
\author{
Danail Obreschkow
}
\keyword{datasets}
