% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readgadget.R
\name{readgadget}
\alias{readgadget}
\title{Read Gadget Data}
\usage{
readgadget(file, type = "auto")
}
\arguments{
\item{file}{filename of file to load}

\item{type}{character specifying the data format. Must be either of: \code{bin} for binary format, \code{hdf} for HDF5 format or \code{auto} to automatically determine the format from file extension.}
}
\value{
Returns a list containing the particle data. The format of the list depends on the input format.
}
\description{
Reads astrophysical N-body data output by the Gadget code
}
\author{
Danail Obreschkow
}
