% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape.R
\name{shape}
\alias{shape}
\title{Moment difference of two point sets}
\usage{
shape(x, y)
}
\arguments{
\item{x}{n-by-3 matrix containing the first set of points (e.g. dark matter)}

\item{y}{m-by-3 matrix containing the second set of points (e.g. gas)}
}
\description{
Evaluates the multipole moment differences of two 3D point sets; designed to study large-scale differences between dark matter and gas in N-body simulations.
}
\examples{
# Generate a mock bullet-cluster and output its quadrupole index
x.gas = cooltools::runif3(4e3)
x.cdm = t(cbind(t(cooltools::runif3(5e3))-c(1,0,0),t(cooltools::runif3(5e3))+c(1,0,0)))
cat(sprintf('Quadrupole index = \%.2f\n',shape(x.cdm,x.gas)$mu[3]))
sphview(rbind(x.gas,x.cdm), c(rep(1,4e3),rep(2,1e4)))

}
\seealso{
\code{\link{sphview4}}
}
\author{
Danail Obreschkow
}
