% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sphview.R
\name{sphview}
\alias{sphview}
\title{Visualise one or seveal 3D point sets}
\usage{
sphview(x, col = c("#ff0010", "#0515ff", "green", "orange", "yellow",
  "purple"), center = NULL, radius = NULL, radius.scale = 1,
  screen = TRUE, pngfile = NULL, pdffile = NULL, rotation = 1,
  kde = TRUE, ngrid = 300, lum = 1, shadows = 1, sigma = 1,
  arrows = TRUE, side = NULL, xlab = NULL, ylab = NULL,
  add = FALSE, add.xlim = NULL, add.ylim = NULL)
}
\arguments{
\item{x}{list of n-by-3 matrices, represending the coordinates of n 3D positions; n can be different for each matrix. The particles stored in different matrices represent different types of particles and are visualised in different colors.}

\item{col}{vector of colors used for the different particle types, corresponding to the matrix-valued elements in the list \code{x}.}

\item{center}{3-vector specifying the coordinate at the center of the plot. The default is the center of mass.}

\item{radius}{radius (= half horizontal/vertical diameter) of the image in the same length units as \code{x}. The default is the largest distance between the center of mass and all points, multiplied by \code{radius.scale}.}

\item{radius.scale}{scaling factor of the \code{radius}, if the latter is not specified by the user.}

\item{screen}{logical flag specifying whether the images is displayed on the screen.}

\item{pngfile}{optional png-filename to save the image as raster image.}

\item{pdffile}{optional pdf-filename to save the image as pdf-file.}

\item{rotation}{either an integer (1-6) or a 3-vector specifying a rotationn of the 3D particle positions. For integer, 1=(x,y)-plane, 2=(y,z)-plane, 3=(x,y)-plane, 4=(qmax,qmin)-plane, 5=(qmax,qmid)-plane, 6=(qmid,qmin)-plane, where qmax/qmid/qmin are the eigenvectors of the particle-quadrupole, associated with the maximum/middle/minimum eigenvalues, respectively. If \code{rotation} is a vector, its direction specified the rotation axis and its norm the rotation angle in the positive geometric sense.}

\item{kde}{logical flag, specifying whether the particles are smoothed using an addaptive kernel density estimator.}

\item{ngrid}{number of grid cells per side in the output image.}

\item{lum}{overall luminosity scaling factor (default 1).}

\item{shadows}{differential luminosity scaling factor for darker regions (default 1).}

\item{sigma}{relative smoothing scale (default 1).}

\item{arrows}{logical flag, specifying if axis-arrows are drawn.}

\item{side}{vector made of up to four integer values (1/2/3/4), specifying which axes with tick marks are to be drawn (see \code{\link{axis}}).}

\item{xlab}{label on horizontal arrow (only shown if \code{arrows=TRUE}).}

\item{ylab}{label on vertical arrow (only shown if \code{arrows=TRUE}).}

\item{add}{logical flag specifying whether the images added to an existing plot. If TRUE, no pdf-file can be produced with this routine.}

\item{add.xlim}{2-vector specifing the horiziontal position (left and right) of the image on the existing plot if \code{add=TRUE}.}

\item{add.ylim}{2-vector specifing the vertical position (bottom and top) of the image on the existing plot if \code{add=TRUE}.}
}
\description{
Produces a square raster image visualising particle positions in 3D N-body/SPH data.
}
\examples{
# Example of 2x1e5 particles representing a homogenous sphere contained in another sphere
x.red = cooltools::runif3(1e5)
x.blue = cooltools::runif3(1e5,r = c(0,0.5))
sphview(list(x.red,x.blue), radius.scale=1.2)

}
\seealso{
\code{\link{sphview4}}
}
\author{
Danail Obreschkow
}
