% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sphview4.R
\name{sphview4}
\alias{sphview4}
\title{Show one or seveal 3D point sets in four projections}
\usage{
sphview4(
  x,
  rotations = c(2, 3, 4, 1),
  screen = TRUE,
  pdffile = NULL,
  title = NULL,
  scale = TRUE,
  ...
)
}
\arguments{
\item{x}{n-by-3 matrix, representing the 3D-coordinates of n particles.}

\item{rotations}{4-vector or 4-element list specifying four different values/vectors of the argument \code{rotation} in \code{\link{sphview}}. These different rotations are shown in the bottom left, bottom right, top left, and top right panel, respectively, in this order.}

\item{screen}{logical flag specifying whether the images is displayed on the screen.}

\item{pdffile}{optional pdf-filename to save the image as pdf-file.}

\item{title}{Text to be added to the figure.}

\item{scale}{logical flag, specifying if a length scale is shown}

\item{...}{additional parameters for \code{\link{sphview}}.}
}
\description{
Produces a 2-by-2 panel figure, where each panel is made of an \code{\link{sphview}} image in a different rotation.
}
\examples{
# Example of a triaxial ellipsoid shown in four projections
x = t(t(cooltools::runif3(1e4))*c(3,2,1))\%*\%cooltools::rotation3(c(1,1,1))
sphview4(x)

}
\seealso{
\code{\link{sphview}}
}
\author{
Danail Obreschkow
}
