% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surfsmovie.R
\name{surfsmovie}
\alias{surfsmovie}
\title{Generate movie for SURFS particles}
\usage{
surfsmovie(file.track, radius = NULL, aspect = 16/9, mp4file,
  fps = 60, H0 = 70, OmegaM = 0.3, OmegaL = 0.7,
  velocity.conversion = 0.00102269032 * (H0/100), dt = 0.05,
  keep.frames = TRUE, rotation = 1, show.time = TRUE,
  text.size = 1, ...)
}
\arguments{
\item{file.track}{filename of the particle tracking file produced by calling "surfsuite trackhalo ..." (see README file of surfsuite).}

\item{radius}{optional value specifying the radius (in simulation units) of the image. For square-movies, this radius is equal to half the horizontal and vertical diameter of the image. Otherwise, it is the geometric mean between these two half-diameters.}

\item{aspect}{aspect ratio of the movie}

\item{mp4file}{filename of mp4-movie}

\item{fps}{frames per second}

\item{H0}{Hubble constant of the simulation in units of km/s/Mpc}

\item{OmegaM}{matter density of the universe at z=0}

\item{OmegaL}{dark energy density of the universe at z=0}

\item{velocity.conversion}{conversion factor from velocity units in the simulation to length units/Gyr (at z=0). The default corresponds to the standard of Gadget-2.}

\item{dt}{time interval between frames in Gyr}

\item{keep.frames}{logical flag specifying whether the individual frame images are to be kept in a temporary folder within the directory specified by \code{\link{paths}}.}

\item{rotation}{either an integer (1-6) or a 3-vector specifying a rotationn of the 3D particle positions. In case of an integer: 1=(x,y)-plane, 2=(y,z)-plane, 3=(x,y)-plane, 4=(qmax,qmin)-plane, 5=(qmax,qmid)-plane, 6=(qmid,qmin)-plane, where qmax/qmid/qmin are the eigenvectors of the particle-quadrupole, associated with the maximum/middle/minimum eigenvalues, respectively. If \code{rotation} is a vector, its direction specifies the rotation axis and its norm the rotation angle in radians in the positive geometric sense.}

\item{show.time}{logical flag specifying whether the lookback time is displayed}

\item{text.size}{scaling factor for the text size used for the lookback time}

\item{...}{additional arguments to passed to \code{\link{sphview}}.}
}
\description{
Turns evolving particle data produced by surfsuite into an mp4 movie. Surfsuite is a fortran tool available at https://github.com/obreschkow/surfsuite. Function requires ffmpeg to be available via the terminal.
}
\seealso{
\code{\link{sphview}} and \code{\link{surfsview}}
}
\author{
Danail Obreschkow
}
