% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surfsread.R
\name{surfsread}
\alias{surfsread}
\title{Load VELOCIraptor groups in SURFS N-body simulation}
\usage{
surfsread(
  haloid = 1,
  subhalos = TRUE,
  snapshot = 199,
  at = NULL,
  parameterset = "L210_N1024-Hydro3D"
)
}
\arguments{
\item{haloid}{halo index in the VELOCIraptor output}

\item{subhalos}{options specifying whether subhalos should be included (only works, if the specified VELOCIraptor output contains substructure)}

\item{snapshot}{simulation snapshot index}

\item{at}{optional snapshot index, specifying at what snapshot the particles of the halo extracted from snapshot "snapshot" should be displayed}

\item{parameterset}{parameterset name; must be matched by a parameterset name in the parameterfile.txt in the path.surfsuite directory. This parameterset name specifies the actual N-body simulation, as well as the VELOCIRAPTOR output.}
}
\description{
Dedicated function to load halos in cosmological N-body simulations, identified by the VELOCIraptor halo-finder. Requires the external tool surfsuite, available at https://github.com/obreschkow/surfsuite.
}
\seealso{
\code{\link{surfsview}} and \code{\link{surfsmovie}}
}
\author{
Danail Obreschkow
}
