% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surfsview.R
\name{surfsview}
\alias{surfsview}
\title{Show VELOCIraptor groups in SURFS N-body simulation}
\usage{
surfsview(
  haloid = 1,
  subhalos = TRUE,
  snapshot = 199,
  at = NULL,
  parameterset = "L210_N1024-Hydro3D",
  species = NULL,
  fourprojections = FALSE,
  ...
)
}
\arguments{
\item{haloid}{halo index in the VELOCIraptor output}

\item{subhalos}{options specifying whether subhalos should be included (only works, if the specified VELOCIraptor output contains substructure)}

\item{snapshot}{simulation snapshot index}

\item{at}{optional snapshot index, specifying at what snapshot the particles of the halo extracted from snapshot "snapshot" should be displayed}

\item{parameterset}{parameterset name; must be matched by a parameterset name in the parameterfile.txt in the path.surfsuite directory. This parameterset name specifies the actual N-body simulation, as well as the VELOCIRAPTOR output.}

\item{species}{optional vector listing the species of particles to be shown}

\item{fourprojections}{logical flag specifying whether the group is visualised in a single projedction (using \code{\link{sphview}}) or in four projections (\code{\link{sphview4}}).}

\item{...}{additional arguments to passed to \code{\link{sphview}} or \code{\link{sphview4}}, respectively.}
}
\description{
Dedicated function to visualise data in ICRAR's SURFS simulation suite, processed by the VELOCIraptor halo-finder. Requires the external tool surfsuite, available at https://github.com/obreschkow/surfsuite.
}
\seealso{
\code{\link{sphview}} and \code{\link{sphview4}}
}
\author{
Danail Obreschkow
}
