% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{paths}
\alias{paths}
\title{Get and set paths to external routines}
\usage{
paths(surfsuite = NULL, procorr = NULL)
}
\arguments{
\item{surfsuite}{path of the tool surfsuite, available at https://github.com/obreschkow/surfsuite}

\item{procorr}{path of the tool procorr, available at https://github.com/obreschkow/procorr}
}
\description{
Get and set paths to external routines
}
\examples{
# Check current paths
paths()

}
\author{
Danail Obreschkow
}
