% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sphview.R
\name{sphview}
\alias{sphview}
\title{Visualise one or several 3D point sets}
\usage{
sphview(
  x,
  species = NULL,
  value = NULL,
  valrange = NULL,
  fix.brightness = TRUE,
  weight = NULL,
  colspecies = c("#ff0010", "#0515ff", "green", "orange", "yellow", "purple"),
  colscale = cooltools::spectrumcolors(1000),
  center = NULL,
  width = NULL,
  aspect = 1,
  thickness = NULL,
  subsampling = 1,
  screen = TRUE,
  pngfile = NULL,
  pdffile = NULL,
  rotation = 1,
  ngrid = 300,
  title = NULL,
  title.origin = NULL,
  lum = 1,
  shadows = 1,
  freeze.lum = FALSE,
  sigma = NULL,
  kde = TRUE,
  arrows = TRUE,
  arrow.origin = NULL,
  arrow.length = NULL,
  arrow.lwd = 1.5,
  scale = TRUE,
  scale.origin = NULL,
  scale.length = NULL,
  scale.lwd = 1.5,
  length.unit = "",
  xlab = NULL,
  ylab = NULL,
  cex = 1,
  text.offset = 0
)
}
\arguments{
\item{x}{n-by-3 matrix, representing the 3D-coordinates of n particles.}

\item{species}{optional argument. If \code{x} is a n-by-3 matrix, \code{species} is a n-element vector giving the integer species indices, which will specify the color of the particles; if not given, all particles will be assumed to be of species=1. If \code{x} is a \code{snapshot} list, \code{species} is a single integer or a vector of integer specifying the species to select; if not given, all species are plotted.}

\item{value}{optional n-element vector specifying a particle property (e.g. temperatures) to be shown in color. This overwrites the colors set by \code{species}.}

\item{valrange}{2-vector specifying the range of values covered by the colors in the vector \code{colorscale}.}

\item{fix.brightness}{logical flag specifying whether the brightness of value-plots should scale with density.}

\item{weight}{optional n-element vector, specifying the weight of different particles.}

\item{colspecies}{vector of colors, specifying the color of the species given in the optional \code{species} vector.}

\item{colscale}{vector of colors used to shown the particle property given by the optional vectors \code{value}.}

\item{center}{optional 3-vector specifying the coordinate at the center of the plot. The default is the geometric center (= center of mass, if all particle masses are equal), if \code{x} is a matrix; and the box center, if \code{x} is a Gadget snapshot list.}

\item{width}{optional horizontal range of the image in the length units of \code{x}. The default corresponds to the full range of particle positions, if \code{x} is a matrix; and to the box size, if \code{x} is a Gadget snapshot list.}

\item{aspect}{aspect ratio (= width/height).}

\item{thickness}{optional thickness in length units of \code{x} of the slice to be plotted. If not given, all particles are shown.}

\item{subsampling}{fraction of particles to be used (if 1, all particles are used, if <1 a random subsample is drawn)}

\item{screen}{logical flag specifying whether the images is displayed on the screen.}

\item{pngfile}{optional png-filename to save the image as raster image.}

\item{pdffile}{optional pdf-filename to save the image as pdf-file.}

\item{rotation}{either an integer (1-6), a 3-vector or a 3-by-3 matrix, specifying a rotation of the 3D particle positions. In case of an integer: 1=(x,y)-plane, 2=(y,z)-plane, 3=(x,y)-plane, 4=(qmax,qmin)-plane, 5=(qmax,qmid)-plane, 6=(qmid,qmin)-plane, where qmax/qmid/qmin are the eigenvectors of the particle-quadrupole, associated with the maximum/middle/minimum eigenvalues, respectively. If \code{rotation} is a vector, its direction specifies the rotation axis and its norm the rotation angle in radians in the positive geometric sense.}

\item{ngrid}{optional number of grid cells per side in the output image. If the image is not square ngrid is interpreted as the geometric mean between the horizontal and the vertical number of pixels, such that the total number of pixels remains about ngrid^2.}

\item{title}{Text to be added to the figure.}

\item{title.origin}{optional 2-vector specifying the position of the title}

\item{lum}{overall luminosity scaling factor (default 1).}

\item{shadows}{differential luminosity scaling factor for darker regions (default 1).}

\item{freeze.lum}{logical flag (default FALSE). If set to TRUE, the luminosity will not be adjusted as a function of the visible particles. This is normally needed to avoid flickering in animations.}

\item{sigma}{smoothing scale in simulation units. The default is NULL, meaning that a value is picked automatically. If a vector is provided, the different elements will be interpreted as corresponding to different species.}

\item{kde}{logical flag, specifying whether the particles are smoothed using an adaptive kernel density estimator. If a vector is provided, the different elements will be interpreted as corresponding to different species.}

\item{arrows}{logical flag, specifying if axis-arrows are drawn}

\item{arrow.origin}{optional 2-vector specifying the bottom left corner of the arrows}

\item{arrow.length}{optional number specifying the length of the arrows}

\item{arrow.lwd}{line width of arrows}

\item{scale}{logical flag, specifying if a length scale is shown}

\item{scale.origin}{optional 2-vector specifying the right end of the length scale}

\item{scale.length}{optional number specifying the length of the length scale (is always rounded to one significant digit)}

\item{scale.lwd}{line width of length scale}

\item{length.unit}{character string of length unit (e.g. "m" or "kpc")}

\item{xlab}{label on horizontal arrow (only shown if \code{arrows=TRUE}).}

\item{ylab}{label on vertical arrow (only shown if \code{arrows=TRUE}).}

\item{cex}{text size}

\item{text.offset}{optional extra offset for the arrow and scale labels}
}
\value{
Returns a list of items
\item{rgb}{ngrid-by-ngrid-by-3 array of containig the rgb colour data between 0 and 1}
\item{radius}{positive real number specifying the radius of the image along the horizontal and vertical axes, in the units of the particle positions}
\item{rotationmatrix}{rotation matrix applied to the input data, i.e. the projection was x = (x\%*\%rot)[,1:2]}
}
\description{
Produces a raster image visualising particle positions in 3D N-body/SPH data.
}
\examples{
# Example of 2x1e4 particles in a triaxial ellipsoid
x.red = t(t(cooltools::runif3(1e4))*c(3,2,1))
x.blue = abs(t(t(cooltools::runif3(1e4,r=c(0.5,1)))*c(3,2,1)))
sphview(rbind(x.red,x.blue), species=c(rep(1,1e4),rep(2,1e4)), center=0)

}
\author{
Danail Obreschkow
}
