% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sphview4.R
\name{sphview4}
\alias{sphview4}
\title{Show one or seveal 3D point sets in four projections}
\usage{
sphview4(x, rotations = c(2, 3, 4, 1), screen = TRUE, pdffile = NULL,
  title = NULL, ...)
}
\arguments{
\item{x}{list of n-by-3 matrices, represending the coordinates of n 3D positions; n can be different for each matrix. The particles stored in different matrices represent different types of particles and are visualised in different colors.}

\item{rotations}{4-vector or 4-list specifying four different values/vectors of the argument \code{rotation} in \code{\link{sphview}}. These different rotations are shown in the bottom left, bottom right, top left, and top right panel, respectively, in this order.}

\item{screen}{logical flag specifying whether the images is displayed on the screen.}

\item{pdffile}{optional pdf-filename to save the image as pdf-file.}

\item{title}{Text to be added to the figure.}

\item{...}{additional parameters for \code{\link{sphview}}.}
}
\description{
Produces a 2-by-2 panel figure, where each panel is made of an \code{\link{sphview}} image in a different rotation.
}
\examples{
# Example of three particle species shown in four projections
x.red = cooltools::runif3(1e5,polarangle = c(0,pi/2), azimuth = c(0,pi*0.75))
x.blue = cooltools::runif3(1e5,polarangle = c(pi/2,pi), azimuth = c(pi*1.25,2*pi))
x.green = t(t(cooltools::runif3(2e4,r=0.5))+c(1,1/4,-1))
sphview4(list(x.blue , x.red, x.green), radius.scale=1.2)

}
\seealso{
\code{\link{sphview}}
}
\author{
Danail Obreschkow
}
